finished part 11

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,14 @@ use DB; //In order to use sql querries
 
 class PostsController extends Controller
 {
+
+
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]); //DR: like this guests can't edit posts
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -77,6 +85,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post= Post::find($id);
+        //check for correct user
+        if(auth()->user()->id!==$post->user_id){
+            return view('posts');
+        }
         return view('posts/edit')->with('post',$post);
     }
 
@@ -110,6 +122,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id!==$post->user_id){
+            return view('posts');
+        }
+
         $post->delete();
         return redirect('/posts');
     }
